Update to v097r17 release.
byuu says:

Changelog:
- ruby: if DirectSoundCreate fails (no sound device present), return
  false from init instead of crashing
- nall: improved edge case return values for
  (basename,pathname,dirname,...)
- nall: renamed file_system_object class to inode
- nall: varuint_t replaced with VariadicNatural; which contains
  .bit,.bits,.byte ala Natural/Integer
- nall: fixed boolean compilation error on Windows
- WS: popa should not restore SP
- GBA: rewrote the CPU/APU cores to use the .bit,.bits functions;
  removed registers.cpp from each

Note that the GBA changes are extremely major. This is about five hours
worth of extremely delicate work. Any slight errors could break
emulation in extremely bad ways. Let's hold off on extensive testing
until the next WIP, after I do the same to the PPU.

So far ... endrift's SOUNDCNT_X I/O test is failing, although that code
didn't change, so clearly I messed up SOUNDCNT_H somehow ...

To compile on Windows:

1. change nall/string/platform.hpp line 47 to

    return slice(result, 0, 3);

2. change ruby/video.wgl.cpp line 72 to

    auto lock(uint32_t*& data, uint& pitch, uint width, uint height) -> bool {

3. add this line to the very top of hiro/windows/header.cpp:

    #define boolean FuckYouMicrosoft

+++ nall/inode.hpp
@@ -8,7 +8,7 @@
 
 namespace nall {
 
-struct file_system_object {
+struct inode {
   enum class time : unsigned { access, modify };
 
   static auto exists(const string& name) -> bool {
@@ -70,8 +70,13 @@ struct file_system_object {
 
   //returns false if 'name' is a directory that is not empty
   static auto remove(const string& name) -> bool {
+    #if defined(PLATFORM_WINDOWS)
+    if(name.endsWith("/")) return _wrmdir(utf16_t(name)) == 0;
+    return _wunlink(utf16_t(name)) == 0;
+    #else
     if(name.endsWith("/")) return rmdir(name) == 0;
     return unlink(name) == 0;
+    #endif
   }
 };
